Move metrics server into control module
The metrics Service implementation that renders prometheus metrics can
be used independently of any specific listener.

This change moves the binding and listening details into the `control`
module, as this seems like the best umbrella for the specifics of
serving things to the control plane.
