+++ src/control/mod.rs
@@ -4,6 +4,8 @@ mod fully_qualified_authority;
 mod observe;
 pub mod pb;
 mod remote_stream;
+mod serve_http;
 
 pub use self::destination::Bind;
 pub use self::observe::Observe;
+pub use self::serve_http::serve_http;
